close #66 テストで各値が安定するように修正

+++ test/dummy/Motor.h
@@ -47,7 +47,7 @@ namespace ev3api {
     void reset();
 
    private:
-    int motorCount;
+    double motorCount;
   };
 }  // namespace ev3api
